A fairly large number of changes and debugs:
Added 'offset' property to the timelapseTraps class. This is an array that tells timelapse how much to offset an image of a particular channel by in order to ensure they are registered properly. 
Also amended timelapseTraps.returnSingleTimePoint and timelapseTraps.AddSecondaryChannel to keep it up to date and to use it when getting loading images. 

Did a little debugging on the timelapseTrapsActiveContour and SegmentConsecutiveTimepoints. It was offsetting the absolute position of cells by a pixel when running the 'ReturnAbsoluteCellCentre' or whatever method. This has been fixed. Also some other minor and major bugs about writing the wrong contour to the data structure in SegmentConsecutiveTimepoints field and getting the wrong number of time points when setting up the timelapseTrapsActiveContour class.

Added some more complicated debug imagining code to PSORadialTimeStack.

Modified findTraps to exclude traps within 'trapImage' of each other. Also cut and pasted most of the code into a probably more useful findTrapsSingleImage function.

Added some new small background function in the ACtiveContourFunctions folder: fairly self explanatory 

Installed deconvolution lab but chose to ignore it for version control. I don't think it's going to be widely used in the future.