Changed the way image transformations are done so that trap handling is now done by a separate set of image transformation files (called Trap…). Probably bad to have it in the same package but not a major issue.
TimelapseTrapsActiveContour now stores the cCellVision file as a property after it is loaded to get trap info. (Just organisational). getTrapInfoFromCellVision was altered to facilitate this.

SegmentConsecutiveTimepoints was changed so that it is now possible to leave the first time point unchanged (handy if you want to correct a time point and then carry on from there).