Made some changes to TimelapseTraps class and methods to add a background image that will correct for the uneven illumination. Changes were made to the ReturnSingleTimepoint and to the addSecondaryTimelapseChannel so that they would multiply the flourescence image by a background correction.
Altered the two timelapse display functions to handle single timepoint timelapses (though with errors).

Added functions in the GeneralFunctions folder to register a set of images based on cross correlation and then to correct for the offset. Became quite an elaborate function with a number of optional inputs. Has been tested adn seems to work to spec.

Small changes to active contour code to change the way traps are handled - changes the way the trapPixelImage is made so that it should now be between 0 and 1. Changed the trap handling function to allow for this. 

Changed the way TimelapseTrapsActiveContour gets the ImageStack so that it can be normalised and changed the image transformation function to do this (hope it will make parameters more robust).

Added to self explanatory general functions.