made code more useable and efficient in a number of ways:
changed CFradialTimeStack so that it does fewer median operations to make it faster. This was done by calculating the change in the cost function parameters due to image median before starting optimisation. Also changed the code so the CFRadialTimeStack does both with and without previous time point, but now via a boolean which indicates whether to ignore the first time point in adding up the image score. Concurrent changes were required in PSORadialTimeSack to accommodate this new way of calling CFRadialTimeStack. 
This makes CFRadialTimeStackWithPT obsolete.

Changed the WriteACResults methods of TimelapseTrapsActiveContour to be able to write a whole lot of cells at once since profiler showed that writing one cell at a time was very inefficient. 

Changed SegmentConsecutiveTimepoints in lots of places to accommodate these new changes to the other methods and functions. Also tried to make the whole function more readable by clearing out unnecessary repetitive bits of code such as treating the first time point separately and handling cells with and without a previous time point separately.