Some minor bug corrections.
Also, addition of the HMM training to identify births for the cExperiment class. This uses the lineageInfo portion of the cExperiment data structure to train a HMM to guess whether a birth event is real or not. This can then be applied to the cExperiment. This seems to reduce the error rate of birth events when compared to a whi5 dataset.