This is primarily a bunch of updates to the fitness extraction and processing data. To improve the fitness calculations the tracking is now much more stringent. The tracking should be done fairly stringently, and then tracklets combined. After tracking and extraction, the daughters need to be classified using a HMM that is trained. As of now there is no way to load a previously trained HMM - to use that to classify the daughter birth probability.
The fitness has been tested using a whi5 dataset of moderate quality. It appears that there is roughly a 10% error in the daughter birth identification ... but error is also due to the poor quality of the whi5 dataset, so the actual error is likely lower than this.

The general function outlier removal removes outliers at each time point based on the percentages you pass it.