changed identify traps of cExperiment to use the first timepoint that you want to process rather than timepoint 1.
changed the track traps so that it reset the 'base image' (one to which it compares to identify drift) when the distance drifted was more than half the trap image size. Helps in very large total drift cases.

Added new feature to cExperiment trap selection method such that you can choose to track the first position and use this to exclude from the auto find traps that would move out of the frame due to drift.