added a date line to active contour method so that you know the time it finished.
various fixes throughout code to get it working when you select a sub set of timepoints to be analysed that doesn't start at 0.

Added new cExperiment methods to add BackgroundCorrection Image and offset image to all cTimelapses in the experiment. Not yet properly GUIfied, not sure how for background correction.