Small changes to the identifyCells code to fix bugs when attempting to use a linear classifier.
Small changes to the return singleTimepoint, to correct for a bug when using the 2x2 binning feature on the camera on Robin.

Created a new image feature set to use on Robin with z-stacks in the bright field. This seems to work quite robustly, but could probably be optimised to reduce the number of features and increase the speed.