Lots of fairly important changes to make the whole active contour system more useable from the GUI and give better results.
Finished parallelising cross correlation and active contour code so that it is now fairly fast.

Changed the way the GUI calls the active contour method so that you have the option of picking either of the active contour methods.

added an editCellVisionTrapOutline so that the trap outline could be improved before traps were selected and active contour method was applied. this changes uses a trap from the current timelapse to get as accurate an outline as possible and sets non trap pixels to zero in the cTrap.trap1/cTrap.trap2 so that the same location os found consistently.

added cells to the ElcoCellVisionTraining that allows a cExperiment file to be easily cut down to turn it into a training set. 

===Bug Fixes ====

changed the way tracking was done so that it initialised trapInfo.

added help lines to a few functions 

minor changes