general bug fixes and small changes stashed to have a clean working directory before possibly making edits to the active contour methods.
changes to extractData and compileData to include those fields Elco wants to use and to allow extraction from only a subset of channels for speed.

Minor changes to the editTrapOutline so that it uses the same active contour methods as the cellVision trainingGUI.

Minor bug fixes for the loading cTimelapses.

Added method to export extracted data to csv file for sharing with people easy like.