Removed the mutual information toolbox mi from the generalfunction/filexchange. Found a better toolbox to use instead. Use the distcorr function for doing correlations if you want good data and results.
Made a lot of small changes to the lineage and mother tracking. This will still need some tweaking, so could introduce a few bugs.

The extractSegArea was changed to do an active contour on the fluorescence channel before using that area to identify the size of the cell. 

The cTrapDisplayProcessing has a section of code just for if you are using robin. This should be changed for other purposes.