Changes to the returnCellMother to add some more parameters.
If you are doing active contour segmentation or FLArea segmentation after you’ve run the normal segmentation mode, it now adds an additional field to each cell with that segmented image. This way, when you extract the data, it should be possible to extract the fluorescence data for channel with whichever segmentation mode you used - and then you can change it without re-running the segmentation. This however, has not been implemented yet, so the default is still the HoughRad segmentation from the original code.

There are also new fields, with the radiusAC (active contour radius) and radiusFL (fluorescent radius), and nucRadius. The nuclear radius is a new function that can be run on channels where there is a fluorescent tag of the nucleus. It will do a sub pixel segmentation of the nucleus, and then determine the equivalent radius. It also determines how for the nucleus is from the centre of the cell as determined by the hough transform. This should be useful for determining births, as the nucleus moves to the edge of the cell when it is budding.

The radius, xloc,bloc, posNum,trapNum, are now only extracted for the first channel. Because these values are identical for all channels, and with the daughter counting resulting in hundreds of thousands of cells, this was changed to reduce the memory overhead.

Also, several bugs in the segment consecutiveTPActiveContour stuff were fixed.