Changes to cellResultsViewingGUI - made changes to GUI just to get it to work. - Also, made a bunch of changes to allow editing of the birth events. Now if you press b, it adds a birth event at the timepjoint that is being displayed. If you press 'x' then it deletes the daughter nearest to the current timepjoint - The new/editted daughters are copied from the HMM birth events, and then edited in a separate  field in lineageInfo.motherInfo.manualInfo.motherLabel - this way the edited daughters can't be accidentally overwritten - Added a function in cExperiment to copy the edited daughter birth times and to overwrite the HMMbirthTimes just in case other code was written to use the HMM birth events
cTimelapse traps - changed how the extractSegAreaFL works. Now if there is now fluorescence image at that timepjoint, instead of just taking the segmented radius and putting that in, it enters a zero in that timepjoint. So, the radiusFL will only be filled with values that come from the FL segmenation. Hopefully this will keep things clearer

sync/Parse log file - Added function to a HOG functions folder that allow you to parse the log file and then sync the times at which the channels were actually acquired. This also added parameter to the cellInf(1) which is syncTimes which is the actual times that the images were acquired. Should be useful for plotting data, especially if the imaging frequency changes.

returnParamRLS - made changes to this so that you can pass it birthTimes instead of using the actual HMMbirth events. Also, it now returns a field that is the parameter you are interested in following the cell cycle start. So, so mean GFP for the 12 tp after the each birth event. Useful to compare how cell cycle proteins change with RLS