A bunch of changes - mostly to try to speed up the cellRes gui so that it isn't too laggy
Also, added a new function when you press 'd' on the keyboard in the gui, it records a death event (displayed on the plot in blue). This can be deleted by pressing 'x' if you are near to it.

Also, added a new function saveTimelapse which saves just the timelapse from the cExperiment to speed up processing time for some of the functions.