Some general bug fixes following testing after merge. Now seems to largely work with some bugs and strangenesses.
modification to saveobj method of time-lapse traps so that things can be removed from the saved object and not for the current instance of the object. Currently used to disentangle the timelapseTrapsActiveContour object.
Also required a copy function for timelapseTrapsActiveContour and some changes to its save function.

Make flat field correction work on stacks.

repairs to centre identification for different cell identification methods (‘whole cell’, ‘whole trap’)

repairs to extraction so it doesn’t change the orientation of trapNum when extracting for a single trap.

cellResultsViewingGUI more robust to changing cells selectable.