Test functions now seem to be working - had to deparallelise.
Some changes so that the logger would still work and ‘shouldLog’ property of experimentTracking behaves as expected.

cleared Omero out of createTimelapsePositions and added to experimentTrackingOmero. 

Some fixes to circshift to make it work on older and newer versions of matlab without complaint.

some changes to suppress ‘Temporary Variable’warnings on segmentACexperimental