continue to clean up cellVision training script.
Added new static method (cellVision.createCellVisionFromExperiment) to select instantiate a cellVision from a cExperiment, allowing the desired scaling to be set and a trap template to be selected. This should prove a useful first step in creating a cellVision model for a new trap type or magnification.

Some minor debugging to make sure scaling worked and to add functionality to existing GUIs.