Refactored curateCellTrackingGUI.
Made broad changes to better include transient cCellVision and cCellMorph properties in code (so now basically never necessary to pass the cellVision and cellMorph objects around explicitly in the functions).

Found strange bug in matlab GUI stuff that effects my curateCellTrackingGUI but not on Julian's computer, so haven't fixed it.