Added a help string to the cExperiment GUI and a push button call back so that it can be displayed when one presses 'h'.
Reorganised the main cExperiment GUI in line with lab discussion.

Changed the way cellAutoSelect worked so that it used a property of cExperiment, and stored said property, for the params. This allows the extraction parameters to be set in the script. This was Ivan't request and is now implemented. The autoselect  from GUI was modifed so that this still brings up the parameter select GUI.

still needs to be more broadly tested.